Convert bootloader to Rust
This regresses to a single stage bootloader, but in the future
the primary bootloader will relocate itself, look in the partition
table to read a volume boot record, which will have its own bootloader
and then begin executing that. The volume bootloader will be multistaged,
allowing for the use of Rust's `libcore` and some rudimentary
filesystem drivers.