Use custom libstd
Provide a partial implementation of libstd, enough for libdisk
to function. Many ugly hacks involved.

Currently not functional. Requires llvm's compiler-rt.a which
seems difficult to build for now.

I should probably credit Mozilla and the Rust Devs for all the code
taken for libstd. TODO if this ever makes it into master.