Remove Rust from bootloader
Rust was tried in the bootloader to experiment with reusing the
filesystem implementation in libdisk which was extracted from the
mkdisk utility. However, this ultimately proved very tedious. Without
this code reuse, the driving motivation for a Rust bootloader is greatly
diminshed.

Additionally, it is difficult to justify the purpose of a Rust
bootloader that contains significant portions of inline assembly. It's
cleaner to just use proper assembly.